Remove fast path from divider.
The divider (from day 0) included a feature where it could cache the last result and avoid the huge latency if it would be repeating the same operation again.
The problem with this is that it adds multiple large comparators to the path that decides if the CPU is going to stall. This is currently the critical path, so removing this is well worth it.